client: use single-precision HUD math

+++ src/client/hud_redraw.cpp
@@ -55,7 +55,7 @@ int CHud :: Redraw( float flTime, int intermission )
 {
 	m_fOldTime = m_flTime;	// save time of previous redraw
 	m_flTime = flTime;
-	m_flTimeDelta = (double)m_flTime - m_fOldTime;
+	m_flTimeDelta = m_flTime - m_fOldTime;
 	
 	// Clock was reset, reset delta
 	if ( m_flTimeDelta < 0 )
